EDITOR-#479 move LEDEffectData to graphql/types
