+++ editor-server/src/db/types/mod.rs
@@ -1,6 +1,5 @@
 //! Data types used in database queries.
 
 pub mod color;
-pub mod led_effect;
 pub mod part;
 pub mod user;
